final updates

+++ tools/remove_user.py
@@ -1,20 +1,21 @@
-# import pathlib
-# import json
 import sys
 sys.path.insert(0, './app/')
 from app import app  # noqa
-from app.models import db, Users, LinkRatings, Ratings  # noqa
-from app.models import AllRatings, AllLinkRatings, Comment  # noqa
+from app.models import db, Users
 
-# print(sys.argv[1])
 username = sys.argv[1]
 user = Users.query.filter_by(username=username).first()
+if user is not None:
+    print()
+    confirm = input(f'Are you sure you want to delete user {username} and all their ratings from database <{app.config["SQLALCHEMY_DATABASE_URI"]}>? (y/n): ')
+    print()
+    if confirm == 'y' or confirm == 'Y':
+        db.session.delete(user)
+        db.session.commit()
+        print(f'User {username} deleted.')
+    else:
+        print(f'User {username} not deleted.')
+else:
+    print(f'No user {username} found.')
 
-# # Remove link ratings
-# user.all_link_ratings.delete()
-
-# print(user.all_link_ratings.all()[1].id)
-
-# user.delete()
-db.session.delete(user)
-db.session.commit()
+print()
